requirements.txt
h11~=0.14.0
pip~=25.0.1
sniffio~=1.3.1
click~=8.1.8
httpcore~=1.0.7
idna~=3.10
certifi~=2025.1.31
anyio~=4.8.0
pytz~=2025.1
httpx~=0.28.1
typing_extensions~=4.12.2
python-telegram-bot
python-dotenv